add notice limit, fix bugs

+++ app/models/user.rb
@@ -148,9 +148,9 @@ class User
 
   def insert_unread_comments_notices!(post)
     c = comments_notices.where( :post_id => post.id )
-    if(c.length > 0)
-      c.destroy
-    end
+    c.destroy if c.length > 0
+    comments_notices.shift if comments_notices.length > 99
+    
     comments_notices << CommentsNotice.new(:post => post)
   end
